moved ProcessGroup object to tests as it is only used there
