+++ planetmint/utils.py
@@ -10,31 +10,6 @@ import multiprocessing
 import setproctitle
 
 
-class ProcessGroup(object):
-    def __init__(self, concurrency=None, group=None, target=None, name=None, args=None, kwargs=None, daemon=None):
-        self.concurrency = concurrency or multiprocessing.cpu_count()
-        self.group = group
-        self.target = target
-        self.name = name
-        self.args = args or ()
-        self.kwargs = kwargs or {}
-        self.daemon = daemon
-        self.processes = []
-
-    def start(self):
-        for i in range(self.concurrency):
-            proc = multiprocessing.Process(
-                group=self.group,
-                target=self.target,
-                name=self.name,
-                args=self.args,
-                kwargs=self.kwargs,
-                daemon=self.daemon,
-            )
-            proc.start()
-            self.processes.append(proc)
-
-
 class Process(multiprocessing.Process):
     """Wrapper around multiprocessing.Process that uses
     setproctitle to set the name of the process when running
